+++ reddwarf/utils/__init__.py
@@ -1,6 +1,6 @@
-from reddwarf.utils.clustering import *
+from reddwarf.utils.clusterer.kmeans import *
 from reddwarf.utils.matrix import *
 from reddwarf.utils.polismath import *
 from reddwarf.utils.stats import *
-from reddwarf.utils.pca import *
+from reddwarf.utils.reducer.pca import *
 from reddwarf.utils.statements import *
